+++ src/components/MapDetails/MapDetails.js
