+++ src/components/MapGallery/MapGallery.js
