+++ src/components/MapThumbnails/MapThumbnails.js
